feat: Add mathfilters app to INSTALLED_APPS
This commit adds the 'mathfilters' app to the INSTALLED_APPS list in the settings.py file. The 'mathfilters' app provides additional template filters for performing mathematical operations in Django templates. This feature enhances the functionality of the social media platform by allowing the use of mathematical operations in template rendering.